+++ CHANGELOG.md
@@ -14,6 +14,20 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 
 ### Fixed
 
+## [2.5.0]
+
+### Added
+
+- timeout of 10 seconds for all image API calls.
+
+### Changed
+
+- minecraft Rcon connection now works using a URI (password@ip:port)
+
+### Fixed
+
+- minor issues
+
 ## [2.4.0]
 
 ### Added